+++ packages/components-color/src/InputColor/utils/color_utils.test.ts
