+++ packages/components-color/src/LuminositySlider/LuminositySlider.test.tsx
