+++ packages/components-theme-editor/package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@looker/components": "^0.9.29",
+    "@looker/components-color": "^0.9.29",
     "@looker/design-tokens": "^0.9.27",
     "lodash": "^4.17.20"
   },
