+++ www/package.json
@@ -4,6 +4,7 @@
   "version": "0.9.29",
   "devDependencies": {
     "@looker/components": "^0.9.29",
+    "@looker/components-color": "^0.9.29",
     "@looker/components-theme-editor": "^0.9.29",
     "@looker/design-tokens": "^0.9.27",
     "@looker/icons": "^0.9.26",
