--- www/src/documentation/components/content/color-wheel.mdx
@@ -1,88 +1,0 @@
----
-title: ColorWheel
-github: Form/Inputs/InputColor/ColorWheel/ColorWheel.tsx
----
-
-The `<ColorWheel/>` renders an HSV based color wheel. The wheel can be rendered at any needed size by providing a `size` property.
-The color wheel defaults to selecting a highly saturated and bright red hue.
-
-[This website has a nice treatment on the HSV color space,](http://learn.leighcotnoir.com/artspeak/elements-color/hue-value-saturation/)
-One takeaway is that HSV color space can be thought of as a cylinder, where each slice of the cylinder is showing a circle at a particular
-value between 0 and 1. `Hue` can be thought of as the angle from the center of the circle can take on values between 0 and 360.
-`Saturation` is the distance from the center on a unit circle.
-
-```jsx
-<ColorWheel size={300} />
-```
-
-## Hue, saturation, and value properties
-
-The `<ColorWheel />` component accepts `hue`, `saturation`, and `value`. Setting the `value` will change the brightness of the
-of the color wheel itself, while changing the `hue` and `saturation` will indicate what color has been selected on the color wheel.
-
-```jsx
-<ColorWheel size={200} hue={140} saturation={0.5} value={1} />
-```
-
-## onColorChanged property
-
-Use this to provide a callback to know when a new (hue, saturation) tuple has been clicked. Note that the brightness will never change for a given
-slice of a color wheel, and should be handled outside the color wheel component.
-
-```jsx
-;() => {
-  const [color, setColor] = useState({
-    h: 140,
-    s: 0.5,
-    v: 0.5,
-  })
-
-  const handleColorStateChange = (hs) => {
-    setColor({
-      ...hs,
-      v: color.v,
-    })
-  }
-
-  const handleSliderChange = (event) => {
-    const value = Number(event.currentTarget.value)
-    setColor({ ...color, v: value / 100 })
-  }
-
-  return (
-    <Flex>
-      <Card raised>
-        <CardContent>
-          <Heading as="h2">Color wheel</Heading>
-          <ColorWheel
-            size={300}
-            hue={color.h}
-            saturation={color.s}
-            value={color.v}
-            onColorChange={handleColorStateChange}
-          />
-
-          <LuminositySlider
-            id="typeinp"
-            min={0}
-            max={100}
-            value={color.v * 100}
-            onChange={handleSliderChange}
-            step={1}
-          />
-        </CardContent>
-      </Card>
-      <Card raised>
-        <CardContent>
-          <Heading as="h2">HSV values</Heading>
-          <List>
-            <ListItem>Hue: {color.h.toFixed(0)}</ListItem>
-            <ListItem>Saturation: {color.s.toFixed(1)}</ListItem>
-            <ListItem>Value: {color.v}</ListItem>
-          </List>
-        </CardContent>
-      </Card>
-    </Flex>
-  )
-}
-```
