--- www/src/documentation/components/content/swatch.mdx
@@ -1,12 +1,0 @@
----
-title: Swatch
-github: Form/Inputs/InputColor/Swatch
----
-
-```jsx
-<Flex justifyContent="space-between">
-  <Swatch />
-  <Swatch color="rgb(79, 42, 186)" />
-  <Swatch color="#FFFF00" />
-</Flex>
-```
