+++ www/src/documentation/components/fields/field-color.mdx
@@ -3,7 +3,7 @@ title: FieldColor
 github: Form/Fields/FieldColor
 ---
 
-This component is composed of an InputColor and a Label. It also has a `ColorWheel` that appears in a `Popover` upon click of the color swatch.
+This component is composed of an InputColor and a Label.
 
 ```jsx
   <FieldColor label="what color?" />
